More progress
We have been starting at the beginning of
~/vivalatable-docs/INSTURCTIONS.md.
We have updated documentation which is not currently inclued in the
repo.
We have fixed stuff. Today in particular was Claudia understanding and
documenting the circles of trust system.
We have brough the events and communities pages back from broken.
This is not a functional app yet.