Progress: Working through the instructions
Claudia is working through the instructions again.
We don't have anything like a full site yet.
We do have several pages with two columns, main column header
navigation, a sidebar with our sidebar navigation that mostly works.
We are moving forward and I think the structure is correct so far.